problem solve : Write a JavaScript program to extract values at specified indexes from a specified array.
Use Array.prototype.filter() and Array.prototype.includes() to pull out the values that are not needed.
Set Array.prototype.length to mutate the passed in an array by resetting its length to 0.
Use Array.prototype.push() to re-populate it with only the pulled values.
Use Array.prototype.push() to keep track of pulled values.